priortity non preemptive CPU scheduling
